Added SAMLServiceProvider.Metadata(), SAMLServiceProvider.SPSigningKeyStore
SAMLServiceProvider.Metadata() dumps SP Metadata (SPSSODescriptor).

It is widely considered poor practice to use the same key for signing
as used for encryption.  IDPs, such as ADFS, allow choosing a different
key for each purpose.  Erroring on the side of caution, provide
SAMLServiceProvider.SPSigningKeyStore to allow the SP to specify an
SP Signing Key that is different than the SP Encryption Key.